Added an option to freeze a card

+++ src/main/java/com/bankproject/bank/repository/AccountsRepository.java
@@ -4,6 +4,6 @@ import com.bankproject.bank.entity.Accounts;
 import com.bankproject.bank.entity.Customers;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountsRepository extends JpaRepository<Accounts, Integer> {
+public interface AccountsRepository extends JpaRepository<Accounts, Long> {
     Accounts findByAccountId(Long id);
 }
