+++ src/main/java/com/bankproject/bank/repository/CustomersRepository.java
@@ -3,7 +3,7 @@ package com.bankproject.bank.repository;
 import com.bankproject.bank.entity.Customers;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomersRepository extends JpaRepository<Customers, Integer> {
+public interface CustomersRepository extends JpaRepository<Customers, Long> {
     Customers findByEmail(String email);
     Customers findByCustomerId(Long id);
